requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.1.0
bcrypt==4.2.0
beanie==1.26.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.1
fastapi-jwt==0.3.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
langchain-openai==0.1.20
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.2
langsmith==0.1.98
lazy-model==0.2.0
marshmallow==3.21.3
motor==3.5.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.40.0
orjson==3.10.6
packaging==24.1
pexpect==4.9.0
pinecone==5.0.0
pinecone-client==5.0.0
pinecone-plugin-inference==1.0.2
pinecone-plugin-interface==0.0.7
ptyprocess==0.7.0
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pymongo==4.8.0
pyngrok==7.2.0
pypdf==4.3.1
pyshorteners==1.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
stripe==10.6.0
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
tqdm==4.66.5
twilio==9.2.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
yarl==1.9.4
